added a seperate user table to be joined in the stream, refactored the code and fixed bugs in the docker build
