+++ .dockerignore
@@ -1,5 +1,4 @@
 .idea
 .venv
-jars/build
 jars/.gradle
 **/__pycache__/
