--- data/file1.csv
@@ -1,7 +1,0 @@
-id,uid,firstname,surname,amount
-1,1,adam,shaw,5
-2,1,adam,shaw,9.99
-3,2,lucy,robinson,0.5
-4,3,john,smith,40
-5,1,adam,shaw,3
-6,3,john,smith,34
