--- data/file2.csv
@@ -1,7 +1,0 @@
-id,uid,firstname,surname,amount
-7,1,adam,shaw,5
-8,1,adam,shaw,9.99
-9,2,lucy,robinson,0.5
-10,3,john,smith,40
-11,1,adam,shaw,3
-12,3,john,smith,34
