+++ data/transactions/file1.csv
@@ -1,0 +1,7 @@
+id,uid,amount
+1,1,5
+2,1,9.99
+3,2,0.5
+4,3,40
+5,1,3
+6,3,34
