+++ data/transactions/file2.csv
@@ -1,0 +1,7 @@
+id,uid,amount
+7,1,5
+8,1,9.99
+9,2,0.5
+10,3,40
+11,1,3
+12,3,34
