+++ data/user/part-1.csv
@@ -1,0 +1,4 @@
+id,firstname,lastname,email,dob
+1,adam,shaw,a.shaw@fake.com,1980-01-01
+2,lucy,robinson,lucy.l.rob@gmail.com,1985-02-02
+3,john,smith,smithyj@hotmail.co.uk,1990-03-03
