+++ docker-compose.yaml
@@ -1,5 +1,3 @@
-version: '2'
-
 services:
   kafka:
     image: 'bitnami/kafka:latest'
