+++ src/consumer_spark.py
@@ -1,29 +1,34 @@
-from src.spark import spark
+from src.utils.spark import spark
 from pyspark.sql import functions as F
 from pyspark.sql import types as T
 
 
-df = (
+user_df = (
+  spark
+  .read
+  .option("header", "true")
+  .csv("data/user")
+)
+
+transaction_df = (
   spark
   .readStream
   .format("kafka")
   .option("kafka.bootstrap.servers", "localhost:9094,localhost:9094")
-  .option("subscribe", "account")
+  .option("subscribe", "transactions")
   .option("startingOffsets", "earliest")
   .load()
 )
 
 # id,uid,firstname,surname,amount
 
-df = (
-  df.selectExpr("CAST(key AS STRING)", "CAST(value AS STRING)")
+transaction_df = (
+  transaction_df.selectExpr("CAST(key AS STRING)", "CAST(value AS STRING)")
   .withColumn("parsed", F.from_json(
     "value",
     schema=T.StructType([
         T.StructField("id", T.StringType()),
         T.StructField("uid", T.StringType()),
-        T.StructField("firstname", T.StringType()),
-        T.StructField("surname", T.StringType()),
         T.StructField("amount", T.StringType()),
     ])
   ))
@@ -32,15 +37,14 @@ df = (
     F.col("value").alias("raw"),
     F.col("parsed.id"),
     F.col("parsed.uid"),
-    F.col("parsed.firstname"),
-    F.col("parsed.surname"),
     F.col("parsed.amount").cast(T.FloatType()).alias("amount")
   )
   .groupBy("uid")
-  .sum("amount")
+  .sum("amount").alias("total")
+  .join(user_df, F.col("id") == F.col("uid"))
 )
 
-query = df.writeStream.outputMode("complete").format("console").start()
+query = transaction_df.writeStream.outputMode("update").format("console").start()
 
 query.awaitTermination()
