+++ src/producer_service.py
@@ -1,5 +1,6 @@
 from confluent_kafka import Producer
-from src.data_loader import get_random_record_json
+from src.utils.data_loader import get_random_record
+import json
 import socket
 import time
 import logging
@@ -29,10 +30,11 @@ def delivery_report(err, msg):
 
 while True:
     try:
-        rec = get_random_record_json("data/file1.csv")
+        rec = get_random_record("data/transactions/file1.csv")
+
         logger.debug("fetched record from producer: %s", rec)
         logger.info("writing record")
-        producer.produce("account", key="key", value=rec, callback=delivery_report)
+        producer.produce("transactions", key=rec["id"], value=json.dumps(rec), callback=delivery_report)
         time.sleep(1)
         producer.flush()
     except Exception as e:
